% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKernalytics.R
\name{kernelLearn}
\alias{kernelLearn}
\title{run Kernalytics in Learn mode}
\usage{
kernelLearn(rootFolder)
}
\arguments{
\item{rootFolder}{folder containg data.csv, desc.csv and algo.csv}
}
\description{
run Kernalytics in Learn mode
}
\author{
Vincent Kubicki
}
