% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runKernalytics.R
\name{kernelPredict}
\alias{kernelPredict}
\title{run Kernalytics in Predict mode}
\usage{
kernelPredict(rootFolder)
}
\arguments{
\item{rootFolder}{folder containg data.csv, desc.csv and algo.csv}
}
\description{
run Kernalytics in Predict mode
}
\author{
Vincent Kubicki
}
