% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offlinechangepoint.R
\name{plot.RealOffCpt}
\alias{plot.RealOffCpt}
\title{Plot signal with cpt}
\usage{
\method{plot}{RealOffCpt}(x, D = NULL, ...)
}
\arguments{
\item{x}{output of \link{realOfflineCpt} function}

\item{D}{number of segments. By default, use the optimal number from slope heuristic}

\item{...}{not used}
}
\description{
Plot signal with cpt
}
\examples{
desc <- matrix(c(0.25, "Gaussian(1.0)", 0.25, "Linear()", 0.5, "Gaussian(0.7)"), nrow = 2)
colnames(desc) = c("GaussA", "GaussA", "GaussB")

data <- data.frame(GaussA = c(rnorm(100, 1, 0.8), rnorm(200, 2, 0.8), rnorm(20, 1, 0.5)),
                   GaussB = c(rnorm(100, 1, 0.8), rnorm(200, 2, 0.8), rnorm(20, 1, 0.5)))

\donttest{
res <- realOfflineCpt(data, desc, Dmax = 6)

plot(res)
}


}
\author{
Quentin Grimonprez
}
