% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offlinechangepoint.R
\name{plotSlopeHeuristic}
\alias{plotSlopeHeuristic}
\title{Plot slope heuristic criterion}
\usage{
plotSlopeHeuristic(res)
}
\arguments{
\item{res}{output of \link{realOfflineCpt} function}
}
\description{
Plot slope heuristic criterion
}
\examples{
desc <- matrix(c(0.25, "Gaussian(1.0)", 0.25, "Linear()", 0.5, "Gaussian(0.7)"), nrow = 2)
colnames(desc) = c("GaussA", "GaussA", "GaussB")

data <- data.frame(GaussA = c(rnorm(100, 1, 0.8), rnorm(200, 2, 0.8), rnorm(20, 1, 0.5)),
                   GaussB = c(rnorm(100, 1, 0.8), rnorm(200, 2, 0.8), rnorm(20, 1, 0.5)))

\dontrun{
res = realOfflineCpt(data, desc, Dmax = 6)

plotSlopeHeuristic(res)
}

}
\author{
Quentin Grimonprez
}
