% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy.R
\name{rScalaBridgeTest}
\alias{rScalaBridgeTest}
\title{rScalaBridgeTest}
\usage{
rScalaBridgeTest(rootFolder)
}
\arguments{
\item{rootFolder}{working directory}
}
\description{
rScalaBridgeTest
}
\author{
Vincent Kubicki
}
