% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offlinechangepoint.R
\name{realOfflineCpt}
\alias{realOfflineCpt}
\title{Offline changepoint segmentation for real data}
\usage{
realOfflineCpt(
  data,
  desc,
  Dmax,
  gramOpti = c("Direct()", "Cache()"),
  folder = "./offlinechangepoint/",
  rmCreatedFiles = TRUE
)
}
\arguments{
\item{data}{data.frame or matrix (do not contain the second line with Real, ...)}

\item{desc}{named matrix containing the kernel to apply}

\item{Dmax}{maximal number of segments}

\item{gramOpti}{méthode LowRank() pas gérée pour le moment}

\item{folder}{folder to save data and output}

\item{rmCreatedFiles}{if TRUE, remove generated files}
}
\value{
a list containing:
\itemize{
  \item D number of segments
  \item tau changepoints (end of segments/ start of segment (-1))
  \item raw.cost
  \item regressed.cost penalty
  \item penalized.cost slope heuristic
  \item data input data
  \item desc kernel description
  \item algo algo description
}
}
\description{
Offline changepoint segmentation for real data
}
\details{
create the following files in folder: c("/algo.csv", "/learnData.csv", "/desc.csv", "/paramTau.csv", "/error.txt")
}
\examples{
desc <- matrix(c(0.25, "Gaussian(1.0)", 0.25, "Linear()", 0.5, "Gaussian(0.7)"), nrow = 2)
colnames(desc) = c("GaussA", "GaussA", "GaussB")

data <- data.frame(GaussA = c(rnorm(100, 1, 0.8), rnorm(200, 2, 0.8), rnorm(20, 1, 0.5)),
                   GaussB = c(rnorm(100, 1, 0.8), rnorm(200, 2, 0.8), rnorm(20, 1, 0.5)))

\donttest{
res <- realOfflineCpt(data, desc, Dmax = 6)
}

}
\author{
Quentin Grimonprez
}
